Refactor metrics calculations for better type handling, update dependencies, and add test dashboard script
- Updated UserMetricsService to ensure avg_matches_per_alert is calculated as a float.
- Updated DeviceTrustService to ensure average_usage is calculated as a float.
- Updated composer.lock to reflect new versions for several Symfony components and other dependencies.
- Added a new public test-dashboard.php script for direct testing of the dashboard bypassing platform checks.
- Minor adjustment in the login form blade template for improved label styling.

+++ app/Services/Analytics/AnalyticsExportService.php
@@ -810,7 +810,7 @@ class AnalyticsExportService
             'platform_url'   => $platform->url,
             'is_active'      => $platform->is_active,
             'total_tickets'  => $platform->scrapedTickets->count(),
-            'avg_price'      => round($platform->scrapedTickets->avg('price') ?: 0, 2),
+            'avg_price'      => round((float) ($platform->scrapedTickets->avg('price') ?: 0), 2),
             'min_price'      => round($platform->scrapedTickets->min('price') ?: 0, 2),
             'max_price'      => round($platform->scrapedTickets->max('price') ?: 0, 2),
             'sports_covered' => $platform->scrapedTickets->pluck('sport')->unique()->count(),
